requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
python-dotenv
openai
tenacity
requests
firebase-admin
b2